RequestScheduleDto : password필드 추가

+++ src/main/java/com/sparta/personaltask2/service/ScheduleService.java
@@ -23,7 +23,6 @@ public class ScheduleService {
         Schedule schedule = new Schedule(requestScheduleDto);
 
         Schedule savedSchedule = scheduleRepository.save(schedule);
-        scheduleRepository.flush();
 
         ResponseScheduleDto responseScheduleDto = new ResponseScheduleDto(schedule);
 
@@ -31,11 +30,16 @@ public class ScheduleService {
     }
 
     public List<ResponseScheduleDto> getSchedules() {
-        return scheduleRepository.findAll().stream().map(ResponseScheduleDto::new).toList();
+        return scheduleRepository.findAllByOrderByCreatedAtDesc().stream().map(ResponseScheduleDto::new).toList();
     }
 
-    public Long deleteSchedule(Long id) {
+    public Long deleteSchedule(Long id,String pw) {
         Schedule schedule = findSchedule(id);
+        if(!isEqualsPassword(pw, schedule.getPassword())) {
+            System.out.println("비밀번호가 일치하지 않습니다");
+            return -1L;
+        }
+
         scheduleRepository.delete(schedule);
         return id;
     }
@@ -43,12 +47,21 @@ public class ScheduleService {
     @Transactional
     public Long updateSchedule(Long id, RequestScheduleDto requestScheduleDto) {
         Schedule schedule = findSchedule(id);
+        if(!isEqualsPassword(requestScheduleDto.getPassword(), schedule.getPassword())) {
+            System.out.println("비밀번호가 일치하지 않습니다");
+            return -1L;
+        }
+
         schedule.update(requestScheduleDto);
 
         return id;
     }
     private Schedule findSchedule(Long id) {
         return scheduleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("선택한 일정은 존재하지 않습니다."));
+    }
+
+    private boolean isEqualsPassword(String pw1, String pw2) {
+        return pw1.equals(pw2);
     }
